+++ scalewiz/components/main_frame.py
@@ -1,4 +1,4 @@
-"""Main Window Tkinter widget for the application"""
+"""Main frame widget for the application."""
 
 # util
 import logging
@@ -6,9 +6,10 @@ from tkinter import ttk
 
 from scalewiz.components.menu_bar import MenuBar
 from scalewiz.components.test_handler_view import TestHandlerView
+from scalewiz.helpers.configuration import get_config
 from scalewiz.models.test_handler import TestHandler
 
-logger = logging.getLogger("scalewiz")
+LOGGER = logging.getLogger("scalewiz")
 
 
 class MainFrame(ttk.Frame):
@@ -36,7 +37,12 @@ class MainFrame(ttk.Frame):
         handler.set_view(view)  # we want to be able to rebuild it later
         self.tab_control.add(view, sticky="nsew")
         self.tab_control.tab(view, text=system_name)
-        logger.info("Added %s to main window", handler.name)
+        LOGGER.info("Added %s to main window", handler.name)
+        # if this is the first handler, open the most recent project
+        if len(self.tab_control.tabs()) == 1:
+            config = get_config()
+            handler.load_project(config["recents"].get("project"))
+            print(handler.project.name.get())
 
     def close(self) -> None:
         """Closes the program if no tests are running."""
@@ -44,7 +50,7 @@ class MainFrame(ttk.Frame):
             widget = self.nametowidget(tab)
             if widget.handler.is_running.get():
                 if not widget.handler.is_done.get():
-                    logger.warning(
+                    LOGGER.warning(
                         "Attempted to close while a test was running on %s",
                         widget.handler.name,
                     )